+++ src/main/java/com/mageddo/dnsproxyserver/di/module/ModuleDao.java
@@ -6,6 +6,8 @@ import com.mageddo.dnsproxyserver.docker.dataprovider.ContainerFacade;
 import com.mageddo.dnsproxyserver.docker.dataprovider.ContainerFacadeDefault;
 import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacade;
 import com.mageddo.dnsproxyserver.docker.dataprovider.DockerNetworkFacadeDefault;
+import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAO;
+import com.mageddo.dnsproxyserver.solver.remote.dataprovider.SolverConsistencyGuaranteeDAOImpl;
 import com.mageddo.dnsproxyserver.solver.docker.dataprovider.ContainerDAO;
 import com.mageddo.dnsproxyserver.solver.docker.dataprovider.ContainerDAODefault;
 import com.mageddo.dnsproxyserver.solver.docker.dataprovider.DockerDAO;
@@ -52,4 +54,7 @@ public interface ModuleDao {
   @Singleton
   DpsContainerDAO dpsContainerDAO(DpsContainerDAODefault impl);
 
+  @Binds
+  @Singleton
+  SolverConsistencyGuaranteeDAO consistencyGuaranteeDAO(SolverConsistencyGuaranteeDAOImpl impl);
 }
