--- src/main/java/com/mageddo/dnsproxyserver/solver/dataprovider/SolverRemoteConsistencyGuaranteeDAO.java
@@ -1,5 +1,0 @@
-package com.mageddo.dnsproxyserver.solver.dataprovider;
-
-public interface SolverRemoteConsistencyGuaranteeDAO {
-  void flushCachesFromCircuitBreakerStateChange();
-}
