+++ src/main/java/com/mageddo/dnsproxyserver/solver/remote/dataprovider/SolverConsistencyGuaranteeDAO.java
@@ -1,0 +1,5 @@
+package com.mageddo.dnsproxyserver.solver.remote.dataprovider;
+
+public interface SolverConsistencyGuaranteeDAO {
+  void flushCachesFromCircuitBreakerStateChange();
+}
